src/requirements.txt
pandas
boto3
line-bot-sdk
